tf_code/requirements.txt
matplotlib
numpy
Pillow
scikit-image
tensorflow
tqdm
pandas
scipy
tensorflow_datasets